+++ livestack/resource-manager-telco/provider.tf
@@ -1,0 +1,8 @@
+provider "oci" {
+  region = var.region
+}
+
+provider "oci" {
+  alias  = "home"
+  region = local.resolved_home_region
+}